Removed IOU
The CEUS consumption numbers include the IOU's and SMUD, which is a POU (Publicly Owned Utility), though not the muni's, the largest of which is LADWP. So, our estimated benefits (bash.69B + bash.064B) include non-IOU buildings, though not all of them.

I suggest we delete IOU in any discussion of benefits to CA and let Ratepayers be a slightly vague generic term.